Changing max number of tasks and semaphores.

+++ arch/software/include/arch_task_context.h
@@ -10,7 +10,7 @@ typedef ucontext_t TaskContext;
 
 // Task critical section state for keeping track of nested call for critical_enter and critical_exit
 typedef struct {
-    uint32_t nesting_count;
+    uint8_t nesting_count;
     sigset_t mask;
 } TaskCritical;
